refactor : modal과 UI 랜더링 코드 분리하기 - 기존의 seminarManage 파일 위치 이동 - 모달 랜더링과 UI 랜더링 부분 분리
#290
